Moves CrossSection and DiffusionCrossSection headers into data folder
