+++ src/scarabee/_scarabee/cross_section.cpp
@@ -1,4 +1,4 @@
-#include <cross_section.hpp>
+#include <data/cross_section.hpp>
 #include <utils/logging.hpp>
 #include <utils/scarabee_exception.hpp>
 #include <utils/constants.hpp>
