+++ src/scarabee/_scarabee/include/data/cross_section.hpp
@@ -1,7 +1,7 @@
 #ifndef SCARABEE_CROSS_SECTIONS_H
 #define SCARABEE_CROSS_SECTIONS_H
 
-#include <diffusion_cross_section.hpp>
+#include <data/diffusion_cross_section.hpp>
 
 #include <xtensor/xtensor.hpp>
 #include <xtensor/xview.hpp>
