+++ src/scarabee/_scarabee/include/moc/empty_cell.hpp
@@ -3,7 +3,7 @@
 
 #include <moc/cell.hpp>
 #include <moc/surface.hpp>
-#include <cross_section.hpp>
+#include <data/cross_section.hpp>
 
 #include <memory>
